requirements.txt
Django==3.2
djangorestframework==3.12.4
django-rest-swagger==2.1.2
pytz==2021.1
